requirements.txt
Pillow>=10.0.0
click>=8.1.0
pathlib2>=2.3.7
tqdm>=4.65.0
pyyaml>=6.0
pytesseract>=0.3.10
pdf2image>=1.16.3
pypdf2>=3.0.1
reportlab>=4.0.4
easyocr>=1.7.0
